BTN API Snatch List Puller
This is a file that pulls the snatch list for up to 4000   but by fidning client.request("GetUserSnatchlist, key=api_key, results= ? You can actually expand beyond 4000.

Essentially it will create a text document called "List of snatched.txt" in the folder it's run under with the Full data from Broadcasthenets API.

This data can then be mined.

yet to be done is to extract out of this information the relevant information  and a script to also extract out the relevant information in your file folders and compare the two.

Co-Authored-By: sorceressofmathematics <sorceressofmathematics@users.noreply.github.com>